requirements.txt
groq==0.5.0
streamlit==1.33.0
sentence-transformers==2.7.0
langchain==0.1.16
langchain-community==0.0.34
langchain-text-splitters==0.0.1
faiss-cpu==1.8.0
openai==1.23.6
llama-index==0.10.33